requirements.txt
airflow-clickhouse-plugin==0.8.1
alembic==1.8.1
anyio==3.6.1
apache-airflow-providers-cncf-kubernetes==3.0.0
apache-airflow-providers-common-sql==1.0.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.1.0
apache-airflow==2.2.5
apispec==3.3.2
argcomplete==2.0.0
attrs==20.3.0
Babel==2.10.3
black==22.6.0
blinker==1.5
cachelib==0.9.0
cachetools==5.2.0
cattrs==1.10.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
clickclick==20.10.2
clickhouse-driver==0.2.4
colorama==0.4.5
colorlog==6.6.0
commonmark==0.9.1
connexion==2.14.0
contextlib2==21.6.0
croniter==1.3.5
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
docutils==0.16
email-validator==1.2.1
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
Flask==1.1.2
google-auth==2.9.1
graphviz==0.20
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
iso8601==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kubernetes==11.0.0
lazy-object-proxy==1.7.1
lockfile==0.12.2
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.0.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
marshmallow==3.17.0
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
platformdirs==2.5.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.1
py==1.11.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-daemon==2.3.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz-deprecation-shim==0.1.0.post0
pytz==2022.1
pytzdata==2020.1
PyYAML==6.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests==2.28.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
schema==0.7.5
setproctitle==1.2.3
setuptools==63.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
SQLAlchemy==1.3.24
swagger-ui-bundle==0.0.9
tabulate==0.8.10
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
unicodecsv==0.14.1
urllib3==1.26.10
websocket-client==1.3.3
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
